% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{assign_rule}
\alias{assign_rule}
\title{Assign Dynamic Treatment Rule from Conditional Average Treatment Effect}
\usage{
assign_rule(data_with_cate, segment_fun = cost_threshold, ...,
  type = c("inferential", "analytic"), unique_rule = TRUE)
}
\arguments{
\item{data_with_cate}{A \code{\link[data.table]{data.table}} containing the
input data, augmented with cross-validated nuisance parameter estimates and
an estimate of the CATE. This input object should b ecreated by successive
calls to \code{\link{set_est_data}} and \code{\link{est_cate}}, or through
a wrapper function that composes these function calls automatically.}

\item{segment_fun}{A particular choice of function for the assignment of a
treatment rule to a segment based on a specified threshold or constraint.
For details on these, consult the documentation in \code{\link{cost_funs}}.}

\item{...}{Additional arguments passed to \code{segment_fun}. For details,
see the documentation for the cost functions in \code{\link{cost_funs}}.}

\item{type}{A \code{character} string (of length one) specifying how the
treatment decision based on the CATE is to be made. There are two options:
\itemize{
\item \code{"inferential"} (the default) uses a hypothesis test to evaluate
whether the estimated CATE is statistically different from a threshold
and assigns a treatment decision based on the resultant p-value.
\item \code{"analytic"} simply evaluates whether the estimated CATE exceeds a
given threshold and assigns treatment to segments for which this holds.
Note that in both cases the threshold must be either provided directly (as
an input to \code{\link{cost_threshold}}) or will be discovered based on
specified constraints (as in \code{\link{cost_budget}}).
}}

\item{unique_rule}{A \code{logical} indicating whether the treatment rule to
be assigned ought to be unique within segments. The default of \code{TRUE}
ensures that all units belonging the the same segment receive an identical
treatment decision. When this is set to \code{FALSE}, subsequent utilities
(e.g., \code{\link{summarize_segments}}) cannot be used to summarize the
assigned treatment rule across segments.}
}
\value{
A \code{\link[data.table]{data.table}} of the full input data
\code{data_cate_est}, augmented with a single additional column \code{rule}
that specifies the treatment decision derived from the estimated CATE and a
strategy specified by the segmentation function \code{segment_fun}.
}
\description{
Perform the assignment of a treatment rule based on the conditional average
treatment effect (CATE) estimate for each individual in the observed data.
The procedure uses a cost function (see \code{\link{cost_funs}}) to decide
whether a given unit should be assigned to the treatment or the control
condition based on the estimated CATE. The treatment decision accommodates
both hard thresholding of the CATE (i.e., analytic evaluation of whether it
exceeds a cutoff) and soft thresholding (i.e., assessing difference of the
CATE estimate from a threshold via a one-sided hypothesis test), with the
latter being the default. Uniqueness of the assigned treatment rule within
segmentation strata is the default, allowing for stratum-specific inference;
however, when such inference is not of interest, this may be disabled.
}
\keyword{internal}
