% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{bound_propensity}
\alias{bound_propensity}
\title{Bounding Propensity Scores}
\usage{
bound_propensity(vals, bounds = c(0.025, 0.975))
}
\arguments{
\item{vals}{A \code{numeric} vector of values in the closed unit interval.}

\item{bounds}{A \code{numeric} vector containing two values, the first being
the minimum allowable value and the second being the maximum allowable for
values appearing in the vector \code{vals} (the previous argument).}
}
\description{
Bounds estimated propensity score values to be within a specified range.
}
\keyword{internal}
