% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{bound_tolerance}
\alias{bound_tolerance}
\title{Bounding to Numerical Tolerance}
\usage{
bound_tolerance(vals, tol = 1e-04)
}
\arguments{
\item{vals}{A \code{numeric} vector of values in the closed unit interval.}

\item{tol}{A \code{numeric} indicating the tolerance limit to which extreme
values should be truncated. Realizations of \code{val} less than \code{tol}
are truncated to \code{tol} while those greater than (1 - \code{tol}) are
truncated to (1 - \code{tol}).}
}
\description{
Bounds extreme values to a specified tolerance level, for use with sensitive
quantities that must be transformed, e.g., via \code{\link[stats]{qlogis}}.
}
\keyword{internal}
