% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{ci_wald}
\alias{ci_wald}
\title{Confidence Intervals for Conditional Effects and Evaluation Measures}
\usage{
ci_wald(estimation_results, segment_by = NULL, param_type = c("cate",
  "effect_measure"), outcome_type = c("continuous", "binary"),
  coverage = getOption("sherlock.ci_covers"))
}
\arguments{
\item{estimation_results}{A \code{\link[data.table]{data.table}} containing
the input data, augmented to include cross-validated nuisance parameter
estimates, an estimate of the CATE, and a treatment rule assigned based on
the estimated CATE via \code{\link{assign_rule}}. This input object should
be generated by successive calls to \code{\link{set_est_data}} and
\code{\link{est_cate}}, or through a wrapper function that composes these
function calls automatically. For cases in which \code{\link{assign_rule}}
has been called prior, any additional information is preserved.}

\item{segment_by}{A \code{character} vector specifying the column names in
\code{data_obs} that correspond to the covariates over which segmentation
should be performed. This should be a strict subset of \code{baseline}.}

\item{param_type}{A \code{character} string (of length one) indicating the
type of estimates for which a Wald-style confidence interval is to be
computed. The choice of \code{"cate"} indicates that this routine is being
invoked thru \code{\link{summarize_segments}} while \code{"effect_measure"}
indicates that the input was generated by \code{\link{est_effect}}.}

\item{outcome_type}{A \code{character} string (of length one) indicating
whether the outcome variable is binary or continuous-valued. For cases in
which the outcome is binary, the confidence interval for the parameter
estimate is constructed on the logit scale and then back-transformed, in
order to respect the fact that the estimate must lie in the unit interval.}

\item{coverage}{A \code{numeric} indicating the nominal rate at which the
Wald-style confidence interval is intended to cover the target parameter.}
}
\value{
A \code{\link[data.table]{data.table}} matching exactly the input
provided in \code{estimation_results}, augmented to include the lower and
upper bounds from the Wald-style confidence intervals constructed.
}
\description{
Wald-style confidence intervals for \code{\link[data.table]{data.table}}s
with additional columns for estimates of the conditional average treatment
effect (CATE) and related segmentation evaluation measures.
}
\keyword{internal}
