% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_process_behavior_chart_limits.R
\name{create_process_behavior_chart_limits}
\alias{create_process_behavior_chart_limits}
\title{Create Process Behavior Chart Limits}
\usage{
create_process_behavior_chart_limits(data, y_var, grouping_var)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{y_var}{Y variable to be plotted on Y axis (required)}

\item{grouping_var}{Variable to group by (optional)}
}
\value{
A tibble with columns created to store values for the moving range, average, UCL and LCL
}
\description{
Creates limits (LCL, UCL) and related calculations (moving range, LCL, UCL) for a Process Behavior Chart and stores them in separate columns
}
