% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_folder.R
\name{create_project_folder}
\alias{create_project_folder}
\title{Create Project Folder}
\usage{
create_project_folder(folder_name, path, subfolders = "generic")
}
\arguments{
\item{folder_name}{Set name of the folder. Examples: "Analysis_20221212", "01_application" (required)}

\item{path}{Set path for folder. Example: "R/Projects/" (required)}

\item{subfolders}{Set type of subfolder structure. Options are "generic" or "shiny". The "generic" option has the following subfolders: "01_data", "02_reports", "03_images", "04_scripts" and "05_misc". The "shiny" option has the following subfolders: "data", "css", "wwww", "images" and "scripts". By default, it is set to "generic". (optional)}
}
\value{
A project folder and sub-folder structure and corresponding .Rproj file on your computer
}
\description{
Creates a project folder on your computer
}
