% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate.R
\name{cv_fit_cate}
\alias{cv_fit_cate}
\title{Cross-Validated Fitting of the Conditional Average Treatment Effect}
\usage{
cv_fit_cate(fold = NULL, data_for_cate, segment_by, cate_learner)
}
\arguments{
\item{fold}{An object specifying the cross-validation folds into which the
observations fall, as generated by \code{\link[origami]{make_folds}}.}

\item{data_for_cate}{A \code{\link[data.table]{data.table}} with additional
columns containing nuisance parameter estimates (propensity score, outcome
regression, doubly robust transformed pseudo-outcome) appended. A dataset
of this form can be used directly for estimation of the CATE.}

\item{segment_by}{A \code{character} vector specifying the column names in
\code{data_obs} that correspond to the covariates over which segmentation
should be performed. This should be a strict subset of \code{baseline}.}

\item{cate_learner}{An instantiated learner object, with class inheriting
from \code{\link[sl3]{Lrnr_base}}, from \pkg{sl3}, to be used to estimate
the CATE, based on a regression of a doubly robust pseudo-outcome on the
specified segmentation covariates. Note that the outcome of this estimation
task is derived from the other nuisance parameter estimates and should be
expected to always be continuous-valued, so algorithms or ensembles should
be set up accordingly.}
}
\value{
A \code{list} (as required by \code{\link[origami]{cross_validate}})
containing a trained \pkg{sl3} learner object in its first slot and the
predicted values of the CATE in its second slot.
}
\description{
Estimation of the conditional average treatment effect (CATE), either within
particular folds induced by cross-validation sample-splitting or upon the
full data. In the latter case, a copy of the full data is made (which is NOT
recommended). Whether the CATE is estimated within folds or not is specified
by the \code{split_type} argument of \code{\link{est_cate}}. Estimation of
the CATE is performed by computing the conditional mean of the doubly robust
transformed pseudo-outcome on the specified set of segmenation covariates.
}
\keyword{internal}
