% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_horizontal_reference_line.R
\name{draw_horizontal_reference_line}
\alias{draw_horizontal_reference_line}
\title{Draw horizontal reference line}
\usage{
draw_horizontal_reference_line(
  reference_line,
  color = "grey",
  linetype = "dashed",
  size = 0.5
)
}
\arguments{
\item{reference_line}{input y coordinate of reference line(s). for multiple reference lines, concatenate individual values into a vector  (required)}

\item{color}{change reference line color. options are "grey", "blue" and "red". by default, it is set to "grey" (optional)}

\item{linetype}{change line type. identical to linetype ggplot2 aesthetic. by default, it is set to "dashed" (optional)}

\item{size}{change line thickness. identical to size ggplot2 aesthetic. by default, it is set to 0.7 (optional)}
}
\value{
A ggplot multi-vari plot object
}
\description{
Draws horizontal reference line(s) to plots
}
\examples{
library(ggh4x)

multi_vari_data \%>\%
    draw_multi_vari_plot(response = force, factor_1 = cycle, factor_2 = fixture, factor_3 = line) +
    draw_horizontal_reference_line(reference_line = c(19.3, 23.5))

}
