% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_interaction_plot.R
\name{draw_interaction_plot}
\alias{draw_interaction_plot}
\title{Draw Interaction Plot}
\usage{
draw_interaction_plot(
  data,
  y_var,
  x_var_1,
  x_var_2,
  alpha = 0.5,
  analysis_desc
)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{y_var}{Y variable to be plotted on Y axis (required)}

\item{x_var_1}{First grouping variable (optional)}

\item{x_var_2}{Second, higher-level grouping variable  (optional)}

\item{alpha}{Set transparency. By default, it is set to 0.5  (optional)}

\item{analysis_desc}{Set whether to add jitter. By default, it is set to TRUE  (optional)}
}
\value{
A ggplot Interaction Plot object
}
\description{
Draws an Interaction Plot
}
