% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_multivari_plot.R
\name{draw_multivari_plot}
\alias{draw_multivari_plot}
\title{Draw Multivari Plot}
\usage{
draw_multivari_plot(
  data,
  response,
  factor_1,
  factor_2,
  factor_3,
  plot_means = FALSE,
  x_axis_text = "normal"
)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{response}{response variable, Y (required)}

\item{factor_1}{lowest level factor (required)}

\item{factor_2}{mid-level factor (required)}

\item{factor_3}{top level factor (optional)}

\item{plot_means}{logical. if FALSE, default, means for mid-level factor are not plotted (optional)}

\item{x_axis_text}{set x axis text size. options are "normal" (default), "small" and "none" (optional)}
}
\value{
A ggplot multi-vari plot object
}
\description{
Draws a multivari small multiples plot
}
\examples{
library(ggh4x)

multivari_data \%>\%
    draw_multivari_plot(response = force, factor_1 = cycle, factor_2 = fixture, factor_3 = line)

multi_vari_data_2 \%>\%
    draw_multivari_plot(response = Length, factor_1 = Trial, factor_2 = Part,
                         factor_3 = Operator, plot_means = TRUE)

}
