% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_multivari_plot_count.R
\name{draw_multivari_plot_count}
\alias{draw_multivari_plot_count}
\title{Draw Multivari Plot for Counts}
\usage{
draw_multivari_plot_count(
  data,
  y_var,
  grouping_var_1,
  grouping_var_2,
  grouping_var_3,
  grouping_var_4,
  x_axis_text_size = 11,
  panel_text_size = 14,
  alpha = 0.6
)
}
\arguments{
\item{data}{Input dataset to be plotted (required)}

\item{y_var}{Response variable, Y (required)}

\item{grouping_var_1}{Select column for lowest level grouping variable (required)}

\item{grouping_var_2}{Select column for second level grouping variable (required)}

\item{grouping_var_3}{Select column for third level grouping variable (optional)}

\item{grouping_var_4}{Select column for fourth level grouping variable (optional)}

\item{x_axis_text_size}{Set X axis text size. By default, it is set to 11  (optional)}

\item{panel_text_size}{Set panel text size. By default, it is set to 14 (optional)}

\item{alpha}{Set transparency. By default, it is set to 0.6  (optional)}
}
\value{
A 'ggplot' object
}
\description{
Draws a multivari small multiples plot for count data
}
