% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pareto_chart.R
\name{draw_pareto_chart}
\alias{draw_pareto_chart}
\title{Draw Pareto Chart}
\usage{
draw_pareto_chart(
  data,
  cat_var,
  y_var,
  highlight_first_n_items = 0,
  lump_last_n_items = 0,
  lumped_cat_name = "Other",
  column_fill = scale_fill_sherlock(3),
  scale = "numeric",
  axis_text_size = 10
)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{cat_var}{categorical variable, Y (required)}

\item{y_var}{(required)}

\item{highlight_first_n_items}{(required)}

\item{lump_last_n_items}{(optional)}

\item{lumped_cat_name}{(optional)}

\item{column_fill}{(optional)}

\item{scale}{(optional)}

\item{axis_text_size}{(optional)}
}
\value{
A Pareto Chart ggplot object
}
\description{
Draws a Pareto Chart
}
