% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_polar_small_multiples.R
\name{draw_polar_small_multiples}
\alias{draw_polar_small_multiples}
\title{Draw Polar Small Multiples}
\usage{
draw_polar_small_multiples(
  data,
  x,
  y,
  grouping_var,
  connect_with_lines = FALSE
)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{x}{x coordinate values (required)}

\item{y}{y coordinate values (required)}

\item{grouping_var}{grouping variable (required)}

\item{connect_with_lines}{logical. if FALSE, default, values within each group are not connected with a line (optional)}
}
\value{
A ggplot polar small multiples object
}
\description{
Draws a Polar Small Multiple Plot
}
