% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_process_behavior_chart.R
\name{draw_process_behavior_chart}
\alias{draw_process_behavior_chart}
\title{Draw Process Behavior Chart}
\usage{
draw_process_behavior_chart(
  data,
  y_var,
  grouping_var,
  limits = TRUE,
  interactive = TRUE
)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{y_var}{Y variable to be plotted on Y axis (required)}

\item{grouping_var}{Variable to group by (optional)}

\item{limits}{Logical. If TRUE, natural process limits (control limits) are plotted. By default, it is set to FALSE  (optional)}

\item{interactive}{Set plot interactivity. By default, it is set to TRUE (optional)}
}
\value{
A 'ggplot' or 'plotly' object
}
\description{
Draws a Process Behavior Chart
}
