% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_process_behavior_chart_limits.R
\name{draw_process_behavior_chart_limits}
\alias{draw_process_behavior_chart_limits}
\title{Draw Process Behavior Chart Limits}
\usage{
draw_process_behavior_chart_limits(
  data,
  x_var,
  avg_line_color = "grey60",
  avg_line_alpha = 0.8,
  avg_line_width = 1,
  limits_color = "red",
  limits_alpha = 0.5,
  limits_width = 1
)
}
\arguments{
\item{data}{Input dataset to be plotted (required)}

\item{x_var}{X variable to be plotted on X axis (required)}

\item{avg_line_color}{Set the color of the grand average line. By default, it is set to "grey60".}

\item{avg_line_alpha}{Set the transparency of the grand average line. By default, it is set to 0.8.}

\item{avg_line_width}{Set the line width of the grand average line. y default, it is set to 1.}

\item{limits_color}{Set the color of the natural process limits (control limits). By default, it is set to "red".}

\item{limits_alpha}{Set the transparency of the natural process limits (control limits). By default, it is set to 0.5.}

\item{limits_width}{Set the line width of the natural process limits (control limits). y default, it is set to 1.}
}
\value{
A ggplot object
}
\description{
Draws limits (LCL, UCL) and grand average for a Process Behavior Chart
}
