% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_small_multiples.R
\name{draw_small_multiples}
\alias{draw_small_multiples}
\title{Draw Small Multiples}
\usage{
draw_small_multiples(
  data,
  x_z_axis_var,
  y_axis_var,
  grouping_var,
  faceting = TRUE,
  unique_color_by_group = TRUE
)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{x_z_axis_var}{variable to be plotted on x and z axes (required)}

\item{y_axis_var}{variable to be plotted on x axis (required)}

\item{grouping_var}{set grouping variable (required)}

\item{faceting}{set whether to display each group in a separate plot. By default, it is set to TRUE (optional)}

\item{unique_color_by_group}{set whether to display each group in a unique color. By default, it is set to TRUE (optional)}
}
\value{
A ggplot Small Multiples object
}
\description{
Draws a Small Multiples Plot
}
