% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_timeseries_scatterplot.R
\name{draw_timeseries_scatterplot}
\alias{draw_timeseries_scatterplot}
\title{Draw Timeseries Scatterplot}
\usage{
draw_timeseries_scatterplot(
  data,
  y_var,
  grouping_var_1,
  grouping_var_1_type = "date-time",
  grouping_var_2,
  faceting = FALSE,
  limits = FALSE,
  date_breaks = "1 month",
  date_labels = "\%b \%y",
  x_axis_text = "normal",
  alpha = 0.3,
  interactive = TRUE
)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{y_var}{Y variable to be plotted on Y axis (required)}

\item{grouping_var_1}{Time variable to be plotted on x axis (required)}

\item{grouping_var_1_type}{Time variable type. Options are "date-time" or "factor"}

\item{grouping_var_2}{Additional variable for faceting (optional)}

\item{faceting}{Set whether to display each group in a separate plot. By default, it is set to FALSE (optional)}

\item{limits}{Logical. If TRUE, process behavior chart control limits for the individual group means are plotted. By default, it is set to FALSE  (optional)}

\item{date_breaks}{Set date breaks. Takes a string, for example "1 week" or "2 days". By default, it is set to "1 month" (optional)}

\item{date_labels}{Set date labels. Identical to the date labels argument of the scale_x_date() ggplot function (optional)}

\item{x_axis_text}{X axis text size. The two options are "normal" and "small" (optional)}

\item{alpha}{Set transparency for the individual observation. Identical to the alpha ggplot argument. By default, it is set to 0.3 (optional)}

\item{interactive}{Set plot interactivity. By default, it is set to TRUE (optional)}
}
\value{
Either a ggplot or plotly Timeseries Scatterplot object
}
\description{
Draws a Timeseries Scatterplot
}
