% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_youden_plot.R
\name{draw_youden_plot}
\alias{draw_youden_plot}
\title{Draw Youden Plot}
\usage{
draw_youden_plot(
  data,
  x_axis_var,
  y_axis_var,
  grouping_var,
  lsl,
  usl,
  median_line = FALSE,
  analysis_desc_label = NULL,
  x_axis_label = "Measurement 1",
  y_axis_label = "Measurement 2"
)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{x_axis_var}{variable to be plotted on x axis (required)}

\item{y_axis_var}{variable to be plotted on x axis (required)}

\item{grouping_var}{grouping variable (optional)}

\item{lsl}{lower specification limit (optional)}

\item{usl}{upper specification limit (optional)}

\item{median_line}{logical. If TRUE, a median bias line is plotted. By default, it is set to FALSE  (optional)}

\item{analysis_desc_label}{Label (subtitle) for analysis description. By default, it is set to NULL  (optional)}

\item{x_axis_label}{Label for x axis. By default, it is set to "Measurement 1"  (optional)}

\item{y_axis_label}{Label for y axis. By default, it is set to "Measurement 2"  (optional)}
}
\value{
A ggplot Youden Plot object
}
\description{
Draws a Youden Plot
}
\examples{
youden_plot_data \%>\%
    draw_youden_plot(x_axis_var = measurement_1, y_axis_var = measurement_2,
                     grouping_var = location)


}
