% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{est_ate}
\alias{est_ate}
\title{Average Treatment Effect of Treating the Optimal Segment}
\usage{
est_ate(data_with_rule, ate_param, est_type)
}
\arguments{
\item{data_with_rule}{A \code{\link[data.table]{data.table}} containing the
input data, augmented with cross-validated nuisance parameter estimates, an
estimate of the CATE, and a treatment rule assigned based on the estimated
CATE via \code{\link{assign_rule}}. This input object should be created by
successive calls to \code{\link{set_est_data}}, \code{\link{est_cate}}, and
\code{\link{assign_rule}} in a sequence, or through a wrapper function that
composes these function calls automatically.}

\item{ate_param}{A \code{character} string (of length one) identifying the
target parameter to be estimated. Choices are \code{"ate_static"} for the
ATE attributable to the static intervention contrasting the two treatment
options at the population level, \code{"ate_dopt_ctl"} for the ATE in the
subgroup identified as potentially being harmed by treatment, as well as
\code{"ate_dopt_trt"} for the ATE in the subgroup identified as benefiting
from treatment.}

\item{est_type}{Specification of either the one-step or TML estimator. See
the documentation of \code{\link{est_effect}} for details.}
}
\value{
A \code{\link[data.table]{data.table}} containing point estimates
and the standard error of the estimates of the specified target parameter.
The resulting object contains the estimated efficient influence function as
an attribute appended to the object via \code{\link[data.table]{setattr}}.
}
\description{
Evaluate the average treatment effect (ATE) based on static interventions or
within subgroups depending on treatment assignment. Efficient one-step and
TML estimators are available for each of these target parameters.
}
\keyword{internal}
