% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{est_effect}
\alias{est_effect}
\title{Efficient Estimation of Causal Effects Under Dynamic Treatment Rules}
\usage{
est_effect(data_with_rule, param_type = c("hte", "ote"),
  est_type = c("onestep", "tmle"))
}
\arguments{
\item{data_with_rule}{A \code{\link[data.table]{data.table}} containing the
input data, augmented with cross-validated nuisance parameter estimates, an
estimate of the CATE, and a treatment rule assigned based on the estimated
CATE via \code{\link{assign_rule}}. This input object should be created by
successive calls to \code{\link{set_est_data}}, \code{\link{est_cate}}, and
\code{\link{assign_rule}} in a sequence, or through a wrapper function that
composes these function calls automatically.}

\item{param_type}{A \code{character} providing a specification for a family
of target parameters to be estimated. The two choices provide estimates of
a range of target parameters. Specifically, the choices correspond to
\enumerate{
\item \code{"hte"}, which computes the average treatment effect (ATE), which
contrasts static interventions for treatment assignment and treatment
being withhold, for (1) the full population, (2) within the subgroup of
units dynamically assigned treatment, and (3) within the subgroup of
units from whom treatment was withheld dynamically. Also included is the
heterogeneous treatment effect (HTE), defined as a difference of the two
subgroup ATEs, which captures the benefit attributable to treating those
units who should receive treatment and withholding treatment from those
that could be harmed by the treatment.
\item \code{"ote"}, which computes several counterfactual means, for (1) the
static intervention of assigning treatment to all units, (2) the static
intervention of withholding treatment from all units, (3) the dynamic
intervention of assigning treatment to those predicted to benefit from
it while withholding treatment from those units that could be harmed. In
addition, two average treatment effects are evaluated, each contrasting
the dynamic treatment rule against the static interventions of assigning
or withholding treatment. The latter three parameters capture contrasts
based on the optimal treatment effect (OTE).
}}

\item{est_type}{A \code{character} specifying the type of estimator to be
computed. Both estimators are asymptotically linear when flexible modeling
choices are used to estimate nuisance parameters, doubly robust (consistent
when at least one nuisance parameter is correctly estimated), and achieve
the best possible variance (i.e., asymptotically efficient) among the class
of regular asymptotically linear estimators. The two options are
\enumerate{
\item \code{"onestep"}, corresponding to the one-step estimator, a first-order
solution to the efficient influence function (EIF) estimating equation.
This is not a substitution (direct) estimator and may be unstable in the
sense of yielding estimates outside the bounds of the target parameter.
\item \code{"tmle"}, corresponding to the targeted minimum loss estimator, an
approach that updates initial estimates of the outcome model by way of a
one-dimensional fluctuation model that aims to approximately solve the
EIF estimating equation.
}}
}
\value{
A \code{\link[data.table]{data.table}} with a handful of rows, one
for each target parameter estimated, and columns giving the parameter name
the point estimate of the target parameter, and the standard error of the
estimate. Also included are the lower and upper confidence limits for the
point estimated based on Wald-style confidence intervals.
}
\description{
Compute efficient estimates of the heterogeneous treatment effect (HTE) or
the optimal treatment effect (OTE), as well as related parameters, including
subgroup average treatment effects and the counterfactual mean of assigning
the dynamic treatment rule.
}
\keyword{internal}
