% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{est_tsm}
\alias{est_tsm}
\title{Efficient Estimation of Counterfactual Means of Dynamic Rules}
\usage{
est_tsm(data_with_rule, tsm_param, est_type)
}
\arguments{
\item{data_with_rule}{A \code{\link[data.table]{data.table}} containing the
input data, augmented with cross-validated nuisance parameter estimates, an
estimate of the CATE, and a treatment rule assigned based on the estimated
CATE via \code{\link{assign_rule}}. This input object should be created by
successive calls to \code{\link{set_est_data}}, \code{\link{est_cate}}, and
\code{\link{assign_rule}} in a sequence, or through a wrapper function that
composes these function calls automatically.}

\item{tsm_param}{A \code{character} string (of length one) identifying the
treatment-specific mean to be estimated. \code{"tsm_static_trt"} gives the
counterfactual mean under the static intervention assigning treatment to
all units; \code{"tsm_static_ctl"} gives the counterfactual mean under the
static intervention withholding treatment from all units; and, lastly,
\code{"tsm_dynamic_opt"} gives the counterfactual mean under the dynamic
rule assigning treatment based on potential benefit/harm from treatment.}

\item{est_type}{Specification of either the one-step or TML estimator. See
the documentation of \code{\link{est_effect}} for details.}
}
\value{
A \code{\link[data.table]{data.table}} containing point estimates
and the standard error of the estimates of the specified target parameter.
The resulting object contains the estimated efficient influence function as
an attribute appended to the object via \code{\link[data.table]{setattr}}.
}
\description{
Evaluate the treatment-specific mean (TSM) based on static interventions or
on the dynamic treatment rule. Efficient one-step and TML estimators are
available for each of these target parameters.
}
\keyword{internal}
