% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance.R
\name{fit_or_mech}
\alias{fit_or_mech}
\title{Outcome Regression Estimation}
\usage{
fit_or_mech(train_data, valid_data, learner)
}
\arguments{
\item{train_data}{A \code{\link[data.table]{data.table}} containing those
observations falling in the training set for a particular cross-validation
sample split. This data object is created by the internal call of
\code{\link{est_cate}} to \code{\link[origami]{cross_validate}} and is
(unfortunately) a copy of a subset of the full estimation data.}

\item{valid_data}{A \code{\link[data.table]{data.table}} containing those
observations falling in the holdout (validation) set for a particular
cross-validation sample split. This data object is created by the internal
call of \code{\link{est_cate}} to \code{\link[origami]{cross_validate}} and
is (unfortunately) a copy of a subset of the full estimation data.}

\item{learner}{An instantiated learner object, with class inheriting from
\code{\link[sl3]{Lrnr_base}}, from \pkg{sl3}, to be used for estimation of
the outcome regression (the mean of the response variable, conditional on
exposure and covariates).}
}
\value{
A \code{list} (as required by \code{\link[origami]{cross_validate}})
containing two slots populated with \code{\link[data.table]{data.table}}s
for the training and validation data (for a given cross-validation split).
Each \code{\link[data.table]{data.table}} has three columns, corresponding
to estimates of the conditional mean of the outcome under the natural value
of the exposure ("QA"), under the control condition ("Q0"), and under the
treatment condition ("Q1").
}
\description{
Estimate the outcome regression (OR), Q(A,W), the conditional mean of the
outcome/response variable, conditional on both the expsoure and the baseline
covariates.
}
\keyword{internal}
