% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance.R
\name{fit_ps_mech}
\alias{fit_ps_mech}
\title{Propensity Score Estimation}
\usage{
fit_ps_mech(train_data, valid_data, learner)
}
\arguments{
\item{train_data}{A \code{\link[data.table]{data.table}} containing those
observations falling in the training set for a particular cross-validation
sample split. This data object is created by the internal call of
\code{\link{est_cate}} to \code{\link[origami]{cross_validate}} and is
(unfortunately) a copy of a subset of the full estimation data.}

\item{valid_data}{A \code{\link[data.table]{data.table}} containing those
observations falling in the holdout (validation) set for a particular
cross-validation sample split. This data object is created by the internal
call of \code{\link{est_cate}} to \code{\link[origami]{cross_validate}} and
is (unfortunately) a copy of a subset of the full estimation data.}

\item{learner}{An instantiated learner object, with class inheriting from
\code{\link[sl3]{Lrnr_base}}, from \pkg{sl3}, to be used for estimation of
the propensity score (the probability of receiving treatment, conditional
on covariates). Note that the outcome of this estimation task is strictly
binary and that algorithms or ensembles should be set up accordingly.}
}
\value{
A \code{list} (as required by \code{\link[origami]{cross_validate}})
containing two slots populated with \code{\link[data.table]{data.table}}s
for the training and validation data (for a given cross-validation split).
Each \code{\link[data.table]{data.table}} has two columns, corresponding to
estimates of the conditional probability of receiving the treatment ("g1")
or the conditional probability of the treatment being withheld ("g0"). Note
that values of these two columns should add up to one for any given row.
}
\description{
Estimate the propensity score (PS), g(1|W), the probability of receiving
treatment conditional on baseline covariates. In a randomized experiment or
A/B test, this is usually known to be g(1|W) = 0.5. In such a case, this
known probability is passed in via \code{\link{Lrnr_prob_known}}.
}
\keyword{internal}
