% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluctuation.R
\name{fluc_update}
\alias{fluc_update}
\title{Fit Fluctuation Model for Targeted Minimum Loss Estimation}
\usage{
fluc_update(data_with_rule, aux_covar, tol = 5)
}
\arguments{
\item{data_with_rule}{A \code{\link[data.table]{data.table}} containing the
input data, augmented with cross-validated nuisance parameter estimates, an
estimate of the CATE, and a treatment rule assigned based on the estimated
CATE via \code{\link{assign_rule}}. This input object should be created by
successive calls to \code{\link{set_est_data}}, \code{\link{est_cate}}, and
\code{\link{assign_rule}} in a sequence, or through a wrapper function that
composes these function calls automatically.}

\item{aux_covar}{A \code{character} string (of length one) indicating the
column of the input data object, \code{data_with_rule}, that corresponds to
the auxiliary ("clever") covariate of the fluctuation model. Note that the
form of this term depends on the efficient influence function of the target
parameter and should be computed externally and appended to the input data.}

\item{tol}{A \code{numeric} indicating the maximum size of the parameter
estimate of the one-dimensional fluctuation model. Any estimates larger
than this tolerance value are set to zero.}
}
\value{
A \code{\link[data.table]{data.table}} with all of the information
contained in the input but augmented to include additional columns for the
updated outcome mechanism based on the one-dimensional fluctuation model.
}
\description{
Fit Fluctuation Model for Targeted Minimum Loss Estimation
}
\keyword{internal}
