% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holmes}
\alias{holmes}
\title{Arthur Conan Doyle's entire Sherlock Holmes works}
\format{
A data frame with 65958 rows and 2 variables: \code{text}
and \code{book}.
}
\source{
\url{https://sherlock-holm.es/}
}
\usage{
holmes
}
\description{
A data set containing the 48 public domain Sherlock Holmes books. This
contains everything except the case-book.
}
\keyword{datasets}
