% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_prob_known.R
\docType{class}
\name{Lrnr_prob_known}
\alias{Lrnr_prob_known}
\title{Predict with Probability Known A Priori}
\format{
An \code{\link[R6]{R6Class}} object inheriting from
\code{\link{Lrnr_base}}.
}
\value{
A learner object inheriting from \code{\link{Lrnr_base}} with
methods for training and prediction. For a full list of learner
functionality, see the complete documentation of \code{\link{Lrnr_base}}.
}
\description{
This learner predicts a binary outcome with a known probability, provided by
the user as input.
}
\section{Parameters}{

\itemize{
\item \code{prob}: The probability of the binary outcome. When not provided,
this defaults to the mean proportion of observed ones.
}
}

\examples{
library(sl3)
data(cpp_imputed)
covs <- c("apgar1", "apgar5", "parity", "gagebrth", "mage", "meducyrs")
task <- sl3_Task$new(cpp_imputed, covariates = covs, outcome = "smoked")

# known probability, as in an RCT
lrnr_rct <- Lrnr_prob_known$new(0.5)
rct_fit <- lrnr_rct$train(task)
rct_preds <- rct_fit$predict()
}
\concept{Learners}
\keyword{data}
