% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_vari_data.R
\docType{data}
\name{multi_vari_data}
\alias{multi_vari_data}
\title{Multi-Vari Plot Sample Dataset 1}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 4 columns.
}
\usage{
multi_vari_data
}
\description{
Contains a sample Multi-Vari Plot dataset
}
\examples{
multi_vari_data

}
\keyword{datasets}
