% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivari_data_2.R
\docType{data}
\name{multi_vari_data_2}
\alias{multi_vari_data_2}
\title{Multi-Vari Plot Sample Dataset 2}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54 rows and 4 columns.
}
\usage{
multi_vari_data_2
}
\description{
Contains a sample Multi-Vari Plot dataset
}
\examples{
multi_vari_data_2

}
\keyword{datasets}
