% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.sherlock}
\alias{plot.sherlock}
\title{Plot Method for Sherlock Summary Output}
\usage{
\method{plot}{sherlock}(x, plot_type = c("cate", "treatment_decisions"),
  sort_by_significance = TRUE, digits = 3L, ...)
}
\arguments{
\item{x}{An object with class \code{sherlock}, which should be a modified
\code{\link[data.table]{data.table}}.}

\item{plot_type}{A \code{character} indicating the type of plot to be made.
Use \code{"cate"} to plot only the CATE values across segments. When
treatment decisions have been assigned by \code{\link{watson_segment}},
then \code{"treatment_decisions"} can be used to plot out the treatment
rules, alongside the CATE estimates.}

\item{sort_by_significance}{A \code{logical}. If \code{TRUE}, the results
will be sorted based on the size of CATE estimates and treatment rules;
if \code{FALSE}, then values will be sorted by the segment dimensions.}

\item{digits}{A \code{numeric} integer giving the number of digits to print
in the entries of the summary table.}

\item{...}{Other options (not currently used).}
}
\value{
ggplot object.
}
\description{
Plot Method for Sherlock Summary Output
}
