% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tukey_duckworth_paired_test.R
\name{plot_tukey_duckworth_paired_test}
\alias{plot_tukey_duckworth_paired_test}
\title{Plot Tukey-Duckworth Paired Test}
\usage{
plot_tukey_duckworth_paired_test(data, y_var, x_vars, arrows = FALSE)
}
\arguments{
\item{data}{input dataset (required)}

\item{y_var}{Y variable of interest (required)}

\item{x_vars}{X variables of interest (required)}

\item{arrows}{Set whether to display arrows in the plot. By default, it is set to FALSE (optional)}
}
\value{
A 'ggplot' object
}
\description{
Plots Tukey-Duckworth Paired Test
}
