% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tukey_duckworth_test.R
\name{plot_tukey_duckworth_test}
\alias{plot_tukey_duckworth_test}
\title{Plot Tukey-Duckworth Test}
\usage{
plot_tukey_duckworth_test(
  data,
  y_var,
  x_var_levels,
  point_size = 3,
  point_type = "solid",
  split_levels = FALSE,
  analysis_desc_label = NULL
)
}
\arguments{
\item{data}{input dataset (required)}

\item{y_var}{Y variable of interest (required)}

\item{x_var_levels}{Levels of X variable of interest (required)}

\item{point_size}{Set point size. By default, it is set to 3. (optional)}

\item{point_type}{Set point size. Options are "solid" (default) and "no fill". (optional)}

\item{split_levels}{Set whether to plot the two levels in separately on the X axis. By default, it is set to FALSE (optional)}

\item{analysis_desc_label}{Label (subtitle) for analysis description. By default, it is set to NULL  (optional)}
}
\value{
A 'ggplot' object
}
\description{
Plots Tukey-Duckworth Paired Test
}
