% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_small_multiples_data.R
\docType{data}
\name{polar_small_multiples_data}
\alias{polar_small_multiples_data}
\title{Polar Small Multiples Sample Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 144 rows and 5 columns.
}
\usage{
polar_small_multiples_data
}
\description{
Contains a sample dataset to demonstrate the use of Polar Small Multiples plot
}
\examples{
polar_small_multiples_data

}
\keyword{datasets}
