% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{print.sherlock}
\alias{print.sherlock}
\title{Print Method for Sherlock Summary Output}
\usage{
\method{print}{sherlock}(x, ..., digits = 4L)
}
\arguments{
\item{x}{An object with class \code{sherlock}, which should be a modified
\code{\link[data.table]{data.table}}.}

\item{...}{Other options (not currently used).}

\item{digits}{A \code{numeric} integer giving the number of digits to print
in the entries of the summary table.}
}
\value{
None. Called for the side effect of printing an informative summary
table from objects of class \code{sherlock}.
}
\description{
Print Method for Sherlock Summary Output
}
