% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{read_table}
\alias{read_table}
\title{Read Table}
\usage{
read_table(path, filetype = ".xlsx")
}
\arguments{
\item{path}{path for the file (required)}

\item{filetype}{set whether to read an .xlsx file or a .csv file. It takes either ".xlsx" or ".csv". By default, it is set to ".xlsx" (optional)}
}
\value{
A tibble
}
\description{
Reads either an .xlsx or a .csv file into a table
}
