% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_analysis.R
\name{save_analysis}
\alias{save_analysis}
\title{Save Analysis}
\usage{
save_analysis(data, plot, filename, filepath)
}
\arguments{
\item{data}{Data to be saved (required)}

\item{plot}{Plot to be saved (optional)}

\item{filename}{Name of the Excel file in a string format without the .xlsx extension. Example: "analysis_results" (required)}

\item{filepath}{Path for the file. Example: "Documents/" (required)}
}
\value{
An Excel file
}
\description{
Saves analysis results, both data and plot, into an .xlsx file
}
