% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_sherlock.R
\name{scale_fill_sherlock}
\alias{scale_fill_sherlock}
\title{Sherlock Fill Color Palettes}
\usage{
scale_fill_sherlock(palette = 1)
}
\arguments{
\item{palette}{fill color palette to be used (required). options are 1, 2 and 3 (2 and 3 are only one color for no grouping). by default it is set to 1.}
}
\value{
A 'ggplot' color scheme that uses one of the Sherlock color fill palettes
}
\description{
Set fill color scheme to one of the Sherlock color palettes
}
