% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_from_unit}
\alias{scale_from_unit}
\title{Scale values from the unit interval to their original scale}
\usage{
scale_from_unit(scaled_vals, max_orig, min_orig)
}
\arguments{
\item{scaled_vals}{A \code{numeric} vector of values scaled to lie in the
closed unit interval by the use of \code{\link{scale_to_unit}}.}

\item{max_orig}{The maximum of the values on the original scale.}

\item{min_orig}{The minimum of the values on the original scale.}
}
\description{
Scale values from the unit interval to their original scale
}
\keyword{internal}
