% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_to_unit}
\alias{scale_to_unit}
\title{Scale values to the unit interval}
\usage{
scale_to_unit(vals)
}
\arguments{
\item{vals}{A \code{numeric} vector of values to be scaled into the closed
unit interval.}
}
\description{
Scale values to the unit interval
}
\keyword{internal}
