% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_low_high_units.R
\name{select_low_high_units}
\alias{select_low_high_units}
\title{Select Low-High Units}
\usage{
select_low_high_units(data, var, number_of_pairs)
}
\arguments{
\item{data}{input dataset (required)}

\item{var}{variable of interest (required)}

\item{number_of_pairs}{Number of low-high pairs to be created. Takes a numeric value (required)}
}
\value{
A tibble object filtered down to the low-high units selected
}
\description{
Automatically selects low-high units in a tibble as well as assigns them into groups
}
