% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_data.R
\name{set_est_data}
\alias{set_est_data}
\title{Inspect the Input Data and Prepare for Estimation Routines}
\usage{
set_est_data(data_obs, baseline, exposure, outcome, segment_by, ids = NULL,
  treatment_cost = NULL)
}
\arguments{
\item{data_obs}{Rectangular input data object, whether a \code{data.frame},
\code{\link[data.table]{data.table}}, or \code{\link[tibble]{tibble}}.}

\item{baseline}{A \code{character} vector specifying the column names in
\code{data_obs} that correspond to the baseline covariates (conditioning
set). These variables should temporally precede the exposure and outcome.}

\item{exposure}{A \code{character} string (of length one) specifying the
column in \code{data_obs} corresponding to the exposure or treatment. This
variable should follow those in \code{baseline} in time but precede the
response variable \code{outcome}.}

\item{outcome}{A \code{character} string (of length one) specifying the
column in \code{data_obs} corresponding to the response variable.}

\item{segment_by}{A \code{character} vector specifying the column names in
\code{data_obs} that correspond to the covariates over which segmentation
should be performed. This should be a strict subset of \code{baseline}.}

\item{ids}{A \code{character} string (of length one) specifying the column
in \code{data_obs} that gives observation-level IDs. The default value of
\code{NULL} assumes that all rows of \code{data_obs} are independent.}

\item{treatment_cost}{A \code{character} string (of length one) specifying the column
in \code{data_obs} that gives observation-level treatment cost. The default value of
\code{NULL} indicates that this is not a cost constrained
problem or that cost is constant for all units.}
}
\description{
A utility function to map input data to abstractions useful for estimation.
This procedure makes a single copy of the input data, storing it as a
\code{\link[data.table]{data.table}} for internal usage in later methods.
}
\keyword{internal}
