% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_multiples_data.R
\docType{data}
\name{small_multiples_data}
\alias{small_multiples_data}
\title{Small Multiples Sample Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2900 rows and 4 columns.
}
\usage{
small_multiples_data
}
\description{
Contains a sample dataset for small multiples
}
\examples{
small_multiples_data

}
\keyword{datasets}
