% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{summarize_segments}
\alias{summarize_segments}
\title{Summarize CATE Estimates and Dynamic Treatment Rule Across Segments}
\usage{
summarize_segments(data_with_rule)
}
\arguments{
\item{data_with_rule}{A \code{\link[data.table]{data.table}} containing the
input data, augmented with cross-validated nuisance parameter estimates, an
estimate of the CATE, and a treatment rule assigned based on the estimated
CATE via \code{\link{assign_rule}}. This input object should be created by
successive calls to \code{\link{set_est_data}}, \code{\link{est_cate}}, and
\code{\link{assign_rule}} in a sequence, or through a wrapper function that
composes these function calls automatically. When \code{\link{assign_rule}}
has not been called, summary of the rule assignment and relevant evaluation
of the CATE is skipped and only segment-specific CATE summaries are output.}
}
\value{
A \code{\link[data.table]{data.table}} containing the discovered
segments, the estimated CATE within each segment (with inference for the
CATE), the segment-specific treatment rule assignment, and auxiliary info
for assessing the quality of the segmentation.
}
\description{
Produce a summary of the segments and the segment-specific treatment rule
assignment based on estimates of the conditional average treatment effect
(CATE) within each segment. When the dynamic treatment rule has not yet been
assigned via \code{\link{assign_rule}}, just segment specific summaries of
the CATE estimates and their standard error are returned.
}
\keyword{internal}
