% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sherlock.R
\name{theme_sherlock}
\alias{theme_sherlock}
\title{Theme Sherlock}
\usage{
theme_sherlock(axis_text_size = "normal")
}
\arguments{
\item{axis_text_size}{set axis text and axis title size. options are "normal" or "small". by default, it is set to "normal"}
}
\value{
A 'theme' object with Sherlock plot theme
}
\description{
Set Sherlock plot theme
}
