% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_scatterplot_data.R
\docType{data}
\name{timeseries_scatterplot_data}
\alias{timeseries_scatterplot_data}
\title{Timeseries Scatterplot Sample Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1170 rows and 5 columns.
}
\usage{
timeseries_scatterplot_data
}
\description{
Contains a sample Timerseries Scatterplot dataset
}
\examples{
timeseries_scatterplot_data

}
\keyword{datasets}
