% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youden_plot_data.R
\docType{data}
\name{youden_plot_data}
\alias{youden_plot_data}
\title{Youden Plot Sample Dataset}
\format{
An object of class \code{data.frame} with 40 rows and 3 columns.
}
\usage{
youden_plot_data
}
\description{
Contains a sample Youden Plot dataset
}
\examples{
youden_plot_data

}
\keyword{datasets}
