% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youden_plot_data_2.R
\docType{data}
\name{youden_plot_data_2}
\alias{youden_plot_data_2}
\title{Youden Plot Sample Dataset 2}
\format{
An object of class \code{data.frame} with 30 rows and 2 columns.
}
\usage{
youden_plot_data_2
}
\description{
Contains a sample Youden Plot dataset
}
\examples{
youden_plot_data_2

}
\keyword{datasets}
