% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cartesian_small_multiples.R
\name{draw_cartesian_small_multiples}
\alias{draw_cartesian_small_multiples}
\title{Draw Cartesian Small Multiple Plot}
\usage{
draw_cartesian_small_multiples(
  data,
  x_coord,
  y_coord,
  grouping_var_1,
  grouping_var_2,
  axes_start_at_zero = FALSE,
  show_axis_values = FALSE,
  faceted = TRUE
)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{x_coord}{x coordinate values (required)}

\item{y_coord}{y coordinate values (required)}

\item{grouping_var_1}{Grouping variable 1 (required)}

\item{grouping_var_2}{Grouping variable 2 (optional)}

\item{axes_start_at_zero}{Logical. Set whether axes start at zero. By default, it is set to FALSE (optional)}

\item{show_axis_values}{Logical. if FALSE, default, axis values are not shown (optional)}

\item{faceted}{logical. if TRUE, default, plot will be faceted. Note: Cartesian plot is always faceted when there are two grouping variables. Drop grouping variable 2 for no faceting. (optional)}
}
\value{
A ggplot cartesian small multiples object
}
\description{
Draws a cartesian small multiples plot
}
