% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cartesian_small_multiples.R
\name{draw_cartesian_small_multiples}
\alias{draw_cartesian_small_multiples}
\title{Draw Cartesian Small Multiple Plot}
\usage{
draw_cartesian_small_multiples(
  data,
  x_coord,
  y_coord,
  grouping_var_1,
  grouping_var_2,
  four_quadrants = FALSE,
  show_axis_values = FALSE,
  faceted = TRUE,
  size = 2,
  alpha = 0.4,
  analysis_desc_label = NULL,
  interactive = FALSE
)
}
\arguments{
\item{data}{Input dataset to be plotted (required)}

\item{x_coord}{x coordinate values (required)}

\item{y_coord}{y coordinate values (required)}

\item{grouping_var_1}{Grouping variable 1 (required)}

\item{grouping_var_2}{Grouping variable 2 (optional)}

\item{four_quadrants}{Logical. Set whether to display four quadrant with both axes starting at zero. By default, it is set to FALSE (optional)}

\item{show_axis_values}{Logical. if FALSE, default, axis values are not shown (optional)}

\item{faceted}{Logical. if TRUE, default, plot will be faceted. Note: Cartesian plot is always faceted when there are two grouping variables. Drop grouping variable 2 for no faceting. (optional)}

\item{size}{Set point size. By default, it is set to 2  (optional)}

\item{alpha}{Set transparency. By default, it is set to 0.4  (optional)}

\item{analysis_desc_label}{Label (subtitle) for analysis description. By default, it is set to NULL  (optional)}

\item{interactive}{Set plot interactivity. By default, it is set to FALSE (optional)}
}
\value{
A 'ggplot' or 'plotly' object
}
\description{
Draws a cartesian small multiples plot
}
