% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_interaction_plot.R
\name{draw_interaction_plot}
\alias{draw_interaction_plot}
\title{Draw Interaction Plot}
\usage{
draw_interaction_plot(
  data,
  y_var,
  x_var_1_levels,
  x_var_2_levels,
  point_size = 4,
  line_size = 1,
  alpha = 0.5,
  analysis_desc_label = NULL
)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{y_var}{Y variable to be plotted on Y axis (required)}

\item{x_var_1_levels}{First grouping variable levels, e.g. -1/1 or "low"/"high" (required)}

\item{x_var_2_levels}{Second grouping variable levels, e.g. -1/1 or "low"/"high" (required)}

\item{point_size}{Set point size. By default, it is set to 4  (optional)}

\item{line_size}{Set line size. By default, it is set to 1  (optional)}

\item{alpha}{Set transparency. By default, it is set to 0.5  (optional)}

\item{analysis_desc_label}{analysis_desc_label Label (subtitle) for analysis description. By default, it is set to NULL  (optional)}
}
\value{
A 'ggplot' object
}
\description{
Draws an Interaction Plot
}
