% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_polar_small_multiples.R
\name{draw_polar_small_multiples}
\alias{draw_polar_small_multiples}
\title{Draw Polar Small Multiples}
\usage{
draw_polar_small_multiples(
  data,
  angular_axis,
  x_y_coord_axis,
  grouping_var,
  faceting_var_1,
  faceting_var_2,
  connect_with_lines = FALSE,
  point_size = 2,
  line_size = 0.6,
  point_alpha = 0.6,
  line_alpha = 0.5,
  label_text_size = 11,
  analysis_desc_label = ""
)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{angular_axis}{angular coordinate values (required)}

\item{x_y_coord_axis}{x-y coordinate values (required)}

\item{grouping_var}{grouping variable (required)}

\item{faceting_var_1}{set first faceting variable (optional)}

\item{faceting_var_2}{set second faceting variable (optional)}

\item{connect_with_lines}{logical. if FALSE, default, values within each group are not connected with a line (optional)}

\item{point_size}{Set point size. By default, it is set to 2  (optional)}

\item{line_size}{Set line size. By default, it is set to 0.6  (optional)}

\item{point_alpha}{Set point transparency. By default, it is set to 0.6  (optional)}

\item{line_alpha}{Set line transparency. By default, it is set to 0.5  (optional)}

\item{label_text_size}{Size of text for labels. By default, it is set to 11  (optional)}

\item{analysis_desc_label}{Label (subtitle) for analysis description. By default, it is set to NULL  (optional)}
}
\value{
A 'ggplot' object
}
\description{
Draws a Polar Small Multiple Plot
}
\examples{
library(dplyr)

polar_small_multiples_data \%>\%
  filter(Mold_Cavity_Number \%in\% c(4, 6)) \%>\%
  draw_polar_small_multiples(angular_axis   = ID_Measurement_Angle,
                             x_y_coord_axis = ID_2,
                             grouping_var   = Tip_Bottom,
                             faceting_var_1 = Mold_Cavity_Number,
                             point_size     = 0.5,
                             connect_with_lines = TRUE,
                             label_text_size = 7)

}
