% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_small_multiples.R
\name{draw_small_multiples}
\alias{draw_small_multiples}
\title{Draw Small Multiples}
\usage{
draw_small_multiples(
  data,
  x_axis_var,
  y_axis_var,
  grouping_var,
  lowest_highest_units,
  faceting = FALSE,
  unique_color_by_group = FALSE,
  interactive = FALSE,
  x_axis_label,
  y_axis_label
)
}
\arguments{
\item{data}{input dataset to be plotted (required)}

\item{x_axis_var}{variable to be plotted on x axis (required)}

\item{y_axis_var}{variable to be plotted on x axis (required)}

\item{grouping_var}{set grouping variable (required)}

\item{lowest_highest_units}{takes a vector of strings corresponding to the lowest/highest units to be highlighted (optional)}

\item{faceting}{set whether to display each group in a separate plot. By default, it is set to FALSE (optional)}

\item{unique_color_by_group}{set whether to display each group in a unique color. By default, it is set to FALSE (optional)}

\item{interactive}{set plot interactivity. By default, it is set to FALSE (optional)}

\item{x_axis_label}{Label for x axis. By default, it is set to "Measurement 1"  (optional)}

\item{y_axis_label}{Label for y axis. By default, it is set to "Measurement 2"  (optional)}
}
\value{
A ggplot Small Multiples object
}
\description{
Draws a Small Multiples Plot
}
