% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file.R
\name{load_file}
\alias{load_file}
\title{Load File}
\usage{
load_file(path, filetype = ".xlsx")
}
\arguments{
\item{path}{path for the file (required)}

\item{filetype}{set whether to read an .xlsx file or a .csv file. It takes either ".xlsx" or ".csv". By default, it is set to ".xlsx" (optional)}
}
\value{
Returns data in the form of a \code{tibble} object.
}
\description{
Reads either an .xlsx or a .csv file into a table
}
