% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivari_data.R
\docType{data}
\name{multi_vari_data}
\alias{multi_vari_data}
\title{Multivari Plot Sample Dataset 1}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 4 columns.
}
\usage{
multi_vari_data
}
\description{
Contains a sample Multivari Plot dataset
}
\examples{
multi_vari_data

}
\keyword{datasets}
