% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_sherlock.R
\name{scale_color_sherlock}
\alias{scale_color_sherlock}
\title{Sherlock Color Palettes}
\usage{
scale_color_sherlock(palette = 1)
}
\arguments{
\item{palette}{color palette to be used (required). options are 1, 2 and 3 (2 and 3 are only one color for no grouping). by default it is set to 1.}
}
\value{
A 'ggplot' color scheme that uses one of the Sherlock color palettes
}
\description{
Set color scheme to one of the Sherlock color palettes
}
